Refactor package structure for user-related entities to enums
